Add all current source files and assets for the ACM Website
The files thus far include one stylesheet, one JavaScript file for
functionality common to the site, image assets that serve as the sites
favicon, logo, and FSU banner, and the core markup files that serve as
the sites general structure at this point in time.
Need to add comments to respective files for maintainability and to ease
contribution process.