Update pages to include a footer, resolves #2
All HTML files now include a footer that uses Font Awesome icons for
club links; their style sheet is referenced in head.
The CSS file has had changes made to the mobile menu to be responsive
for different screen sizes, as well as the footer.
The common.js file has been modified to correct an 'off by one' error in
the calcHeight() function.